fix a bunch of warnings. casts are evil.
git-svn-id: https://svn.codehaus.org/mule/trunk@1498 bf997673-6b11-0410-b953-e057580c5b09

+++ providers/http/src/java/org/mule/providers/http/ResponseWriter.java
@@ -73,7 +73,7 @@ public class ResponseWriter extends FilterWriter {
 
     public void write(byte b) throws IOException {
         super.flush();
-        outStream.write((int)b);
+        outStream.write(b);
     }
     
     public void write(byte[] b) throws IOException {
